code/embedding-search/ingestion/shared/requirements.txt
google-cloud-storage==2.14.0
google-cloud-run==0.10.0
google-cloud-aiplatform==1.38.1
PyMuPDF==1.23.8
Pillow==10.1.0
tiktoken==0.5.2

